# crates/admin-api-client/src/models/error_response.rs
/*
 * MTC-CA Admin API
 *
 * Administrative HTTP+JSON API for the Merkle Tree Certificate CA (spec §17.2). Consumed by `mtcctl` (via the generated Rust client) and the admin UI. Seed surface: health endpoints (§20.5) and a minimal service status endpoint (§17.3).
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

/// ErrorResponse : Standard error envelope for non-2xx responses.
#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct ErrorResponse {
    /// Stable machine-readable error code.
    #[serde(rename = "code")]
    pub code: String,
    /// Human-readable error description.
    #[serde(rename = "message")]
    pub message: String,
}

impl ErrorResponse {
    /// Standard error envelope for non-2xx responses.
    pub fn new(code: String, message: String) -> ErrorResponse {
        ErrorResponse { code, message }
    }
}