# crates/admin-api-client/src/models/readiness_check.rs
/*
 * MTC-CA Admin API
 *
 * Administrative HTTP+JSON API for the Merkle Tree Certificate CA (spec §17.2). Consumed by `mtcctl` (via the generated Rust client) and the admin UI. Seed surface: health endpoints (§20.5) and a minimal service status endpoint (§17.3).
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

/// ReadinessCheck : One dependency check contributing to readiness.
#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct ReadinessCheck {
    /// Check identifier (e.g. `storage`, `hsm`).
    #[serde(rename = "name")]
    pub name: String,
    /// Whether this check passed.
    #[serde(rename = "ok")]
    pub ok: bool,
    /// Human-readable failure detail; omitted when `ok`.
    #[serde(rename = "detail", skip_serializing_if = "Option::is_none")]
    pub detail: Option<String>,
}

impl ReadinessCheck {
    /// One dependency check contributing to readiness.
    pub fn new(name: String, ok: bool) -> ReadinessCheck {
        ReadinessCheck {
            name,
            ok,
            detail: None,
        }
    }
}