# crates/admin-api-client/src/models/readyz_response.rs
/*
 * MTC-CA Admin API
 *
 * Administrative HTTP+JSON API for the Merkle Tree Certificate CA (spec §17.2). Consumed by `mtcctl` (via the generated Rust client) and the admin UI. Seed surface: health endpoints (§20.5) and a minimal service status endpoint (§17.3).
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

/// ReadyzResponse : Readiness probe response (spec §20.5).
#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct ReadyzResponse {
    /// Aggregate readiness of the service.
    #[serde(rename = "status")]
    pub status: Status,
    /// Individual dependency checks backing the aggregate.
    #[serde(rename = "checks")]
    pub checks: Vec<models::ReadinessCheck>,
}

impl ReadyzResponse {
    /// Readiness probe response (spec §20.5).
    pub fn new(status: Status, checks: Vec<models::ReadinessCheck>) -> ReadyzResponse {
        ReadyzResponse { status, checks }
    }
}
/// Aggregate readiness of the service.
#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
pub enum Status {
    #[serde(rename = "ready")]
    Ready,
    #[serde(rename = "not_ready")]
    NotReady,
}

impl Default for Status {
    fn default() -> Status {
        Self::Ready
    }
}