# crates/admin-api-client/src/models/service_info.rs
/*
 * MTC-CA Admin API
 *
 * Administrative HTTP+JSON API for the Merkle Tree Certificate CA (spec §17.2). Consumed by `mtcctl` (via the generated Rust client) and the admin UI. Seed surface: health endpoints (§20.5) and a minimal service status endpoint (§17.3).
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

/// ServiceInfo : Identity and build information for the responding instance.
#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct ServiceInfo {
    /// Service name.
    #[serde(rename = "name")]
    pub name: String,
    /// Semantic version of the running build.
    #[serde(rename = "version")]
    pub version: String,
    /// Cloud region (or simulated region) of this instance.
    #[serde(rename = "region")]
    pub region: String,
    /// Instance start time (RFC 3339).
    #[serde(rename = "started_at", skip_serializing_if = "Option::is_none")]
    pub started_at: Option<chrono::DateTime<chrono::FixedOffset>>,
}

impl ServiceInfo {
    /// Identity and build information for the responding instance.
    pub fn new(name: String, version: String, region: String) -> ServiceInfo {
        ServiceInfo {
            name,
            version,
            region,
            started_at: None,
        }
    }
}